fix: move types of out of the types file
