+++ src/lib/hooks.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from "react";
 import { useQuery } from "react-query";
 import { getAuthToken, getPeopleAsCSVFromAPI } from "./requests";
-import { WonderwallJwtPayload } from "./types";
+import { WonderwallJwtPayload } from "./tokenValidation";
 import { csvToArrayOfObjects, parseJwt } from "./utils";
 
 /**
