+++ src/lib/tokenValidation.ts
@@ -3,8 +3,27 @@ import jwkToPem from "jwk-to-pem";
 import { authConfig } from "./constants";
 import logger from "./logger";
 import { getAzureAdConfig, getPublicJwk } from "./requests";
-import { WonderwallJwtPayload } from "./types";
 
+export type WonderwallJwtPayload = {
+  aud: string;
+  iss: string;
+  iat: number;
+  nbf: number;
+  exp: number;
+  aio: string;
+  azp: string;
+  azpacr: string;
+  groups: string[];
+  name: string;
+  oid: string;
+  preferred_username: string;
+  rh: string;
+  scp: string;
+  sub: string;
+  tid: string;
+  uti: string;
+  ver: string;
+};
 function isValidGroupsClaim(decodedToken: WonderwallJwtPayload | null): boolean {
   return (
     decodedToken?.groups.some(
