+++ src/lib/utils.ts
@@ -1,5 +1,5 @@
 import Papa from "papaparse";
-import { WonderwallJwtPayload } from "./types";
+import { WonderwallJwtPayload } from "./tokenValidation";
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export function debounce<T extends (...argsx: any[]) => void>(callback: T, delay = 500) {
   let timeout: NodeJS.Timeout;
